+++ src/routes/graphql/mutation/changeUser/changeUser.ts
@@ -1,5 +1,5 @@
 import { GraphQLNonNull } from 'graphql';
-import { userObjectType } from '../../queryTypes/userQuery/userObjectType.js';
+import { userObjectType } from '../../query/userQuery/userObjectType.js';
 import { Context } from '../../types/context.js';
 import { UUIDType } from '../../types/uuid.js';
 import { changeUserObjectType } from './changeUserObjectType.js';
