+++ src/routes/graphql/mutation/deletePost/deletePost.ts
