+++ src/routes/graphql/mutation/deleteProfile/deleteProfile.ts
