+++ src/routes/graphql/mutation/deleteUser/deleteUser.ts
