+++ src/routes/graphql/query/memberQuery/memberObjectType.ts
